Include .h instead of .cpp when possible
* Move implementation code from Arduino.h

+++ drivers/Linux/EthernetClient.cpp
@@ -1,24 +1,25 @@
 /*
-* The MySensors Arduino library handles the wireless radio link and protocol
-* between your home built sensors/actuators and HA controller of choice.
-* The sensors forms a self healing radio network with optional repeaters. Each
-* repeater and gateway builds a routing tables in EEPROM which keeps track of the
-* network topology allowing messages to be routed to nodes.
-*
-* Created by Marcelo Aquino <marceloaqno@gmail.org>
-* Copyright (C) 2016 Marcelo Aquino
-* Full contributor list: https://github.com/mysensors/MySensors/graphs/contributors
-*
-* Documentation: http://www.mysensors.org
-* Support Forum: http://forum.mysensors.org
-*
-* This program is free software; you can redistribute it and/or
-* modify it under the terms of the GNU General Public License
-* version 2 as published by the Free Software Foundation.
-*
-* Based on Arduino ethernet library, Copyright (c) 2010 Arduino LLC. All right reserved.
-*/
+ * The MySensors Arduino library handles the wireless radio link and protocol
+ * between your home built sensors/actuators and HA controller of choice.
+ * The sensors forms a self healing radio network with optional repeaters. Each
+ * repeater and gateway builds a routing tables in EEPROM which keeps track of the
+ * network topology allowing messages to be routed to nodes.
+ *
+ * Created by Marcelo Aquino <marceloaqno@gmail.org>
+ * Copyright (C) 2016 Marcelo Aquino
+ * Full contributor list: https://github.com/mysensors/MySensors/graphs/contributors
+ *
+ * Documentation: http://www.mysensors.org
+ * Support Forum: http://forum.mysensors.org
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * version 2 as published by the Free Software Foundation.
+ *
+ * Based on Arduino ethernet library, Copyright (c) 2010 Arduino LLC. All right reserved.
+ */
 
+#include <cstdio>
 #include <sys/socket.h>
 #include <sys/ioctl.h>
 #include <netdb.h>
@@ -249,6 +250,6 @@ bool EthernetClient::operator==(const EthernetClient& rhs) {
 	return _sock == rhs._sock && _sock != -1 && rhs._sock != -1;
 }
 
-uint8_t EthernetClient::getSocketNumber() {
+int EthernetClient::getSocketNumber() {
   return _sock;
 }
